Made quite a few documentation changes
