+++ packages/identityprovider/src/authentication/get-auth-session/get-auth-session.spec.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosResponse } from "axios";
-import { AuthSession, getAuthSession, UnauthorizedError } from "../index";
+import { AuthSession, getAuthSession, UnauthorizedError } from "../../index";
 
 jest.mock("axios");
