+++ packages/identityprovider/src/authentication/get-auth-session/get-auth-session.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosResponse } from "axios";
-import { AuthSession, UnauthorizedError } from "../index";
+import { AuthSession, UnauthorizedError } from "../../index";
 
 export interface AuthSessionDto {
   AuthSessionId: string;
@@ -7,22 +7,21 @@ export interface AuthSessionDto {
 }
 
 /**
- * Returns an [AuthSession]{@link AuthSession} based on an API-Key.
- *
+ * Returns an [AuthSession]{@link AuthSession} based on an API-Key. API-Keys can be generated by d.velop cloud tenant administrators.
  * **The AuthSessionId should be kept secret and never be publicly available.**
  *
  * @param {string} systemBaseUri SystemBaseUri for the tenant
  * @param {string} apiKey API-Key provided by the d.velop cloud
- * @returns {string}
- *
  * @throws {@link UnauthorizedError} indicates an invalid authSessionId or no authSessionId was sent.
  *
- * @example ```typescript
+ * @category Authentication
  *
+ * @example ```typescript
  * const authSession: AuthSession = await getAuthSession("https://monster-ag.d-velop.cloud", "<API_KEY>");
  * console.log(authSession.id); //a valid authSessionId
  * console.log('still valid:', authSession.expire.getTime() > new Date().getTime()); //still valid: true
  * ```
+ *
  */
 export async function getAuthSession(systemBaseUri: string, apiKey: string): Promise<AuthSession> {
