+++ packages/identityprovider/src/authentication/get-impersonated-auth-session-id/get-impersonated-auth-session-id.spec.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosResponse } from "axios";
-import { getImpersonatedAuthSessionId, UnauthorizedError } from "../index";
+import { getImpersonatedAuthSessionId, UnauthorizedError } from "../../index";
 
 jest.mock("axios");
