+++ packages/identityprovider/src/authentication/get-impersonated-auth-session-id/get-impersonated-auth-session-id.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { UnauthorizedError } from "../index";
+import { UnauthorizedError } from "../../index";
 
 interface ImpersonatedAuthSessionDto {
   authSessionId: string;
@@ -11,9 +11,14 @@ interface ImpersonatedAuthSessionDto {
  * @param {string} systemBaseUri SystemBaseUri for the tenant
  * @param {string} appSession A valid appSession
  * @param {string} userId User that should be impersonated
- * @returns Valid authSessionId for the given user
  *
- * //TODO: example
+ * @category Authentication
+ *
+ * @example ```typescript
+ * const authSessionIdForSulley = await getImpersonedAuthSessionId("https://monster-ag.d-velop.cloud", "<APP_SESSION>", "<SULLEYS_USER_ID>");
+ * const sulley: ScimUser = await validateAuthSessionId("https://monster-ag.d-velop.cloud", authSessionIdForSulley);
+ * console.log(sulley.displayName) // James P. Sullivan
+ * ```
  */
 export async function getImpersonatedAuthSessionId(systemBaseUri: string, appSession: string, userId: string): Promise<string> {
