+++ packages/identityprovider/src/authentication/get-login-redirection-uri/get-login-redirection-uri.spec.ts
@@ -1,4 +1,4 @@
-import { getLoginRedirectionUri } from "../index";
+import { getLoginRedirectionUri } from "../../index";
 
 describe("getLoginUri", () => {
   [
