+++ packages/identityprovider/src/authentication/get-login-redirection-uri/get-login-redirection-uri.ts
@@ -2,10 +2,11 @@
  * Returns the URI a request should be redirected to in case it does not contain a valid authSessionId. Redirected requests will display a HTML Login for users to authenticate.
  *
  * @param {string} successUri The URI the request should be redirected to after a successful login
- * @returns {string}
+ *
+ * @category Authentication
  *
  * @example ```typescript
- * const redirectionUri = getLoginRedirectionUri('https://tenant.d-velop.cloud/acme-app/success');
+ * const redirectionUri = getLoginRedirectionUri("<ORIGINAL_URL>");
  * request.redirect(redirectionUri);
  * ```
  */
