+++ packages/identityprovider/src/authentication/request-app-session/request-app-session.spec.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { requestAppSession } from "../index";
+import { requestAppSession } from "../../index";
 
 jest.mock("axios");
