+++ packages/identityprovider/src/authentication/validate-app-session-signature/validate-app-session-signature.spec.ts
@@ -1,4 +1,4 @@
-import { validateAppSessionSignature, InvalidAppSessionSignatureError } from "../index";
+import { validateAppSessionSignature, InvalidAppSessionSignatureError } from "../../index";
 
 describe("validateAppSessionSignature", () => {
