+++ packages/identityprovider/src/authentication/validate-app-session-signature/validate-app-session-signature.ts
@@ -1,18 +1,27 @@
 import { createHash } from "crypto";
-import { InvalidAppSessionSignatureError } from "../index";
+import { InvalidAppSessionSignatureError } from "../../index";
 
 /**
- * Validate the sign value which is provided when an appSession is sent.
+ * Validate the sign value which is provided when an appSession is sent. For further information on this process refer to the [documentation](https://developer.d-velop.de/documentation/idpapi/en/identityprovider-app-201523580.html#IdentityproviderApp-Inter-appcommunicationwithappsessions).
  *
  * @param {string} appName App that requested the appSession
  * @param {string} requestId The requestId that was sent when requesting an appSession
  * @param {string} authSessionId The authSessionId provided to your callback
  * @param {string} expire The expire value provided to your callback
  * @param {string} sign The sign value provided to your callback
- *
  * @throws {@link InvalidAppSessionSignatureError} indicates that the sign value is not valid
  *
- * //TODO: example
+ * @category Authentication
+ *
+ * @example ```typescript
+ * acceptPostRequest(req: Request, res: Respone) {
+ *   try {
+ *     validateAppSessionSignature("acme-app", "requestIdUsedOnAppSessionRequest", req.body.authSessionId, req.body.expire, req.body.sign); //pass or error
+ *   } catch(e) {
+ *     res.send("Forbidden");
+ *   }
+ * }
+ * ```
  */
 export function validateAppSessionSignature(appName: string, requestId: string, authSessionId: string, expire: string, sign: string): void {
   const expectedSign: string = createHash("sha256").update(appName + authSessionId + expire + requestId, "utf8").digest("hex");
