+++ packages/identityprovider/src/authentication/validate-auth-session-id/validate-auth-session-id.ts
@@ -1,16 +1,16 @@
 import axios, { AxiosResponse } from "axios";
-import { ScimUser, ForbiddenError, NotFoundError, UnauthorizedError } from "../index";
+import { ScimUser, ForbiddenError, NotFoundError, UnauthorizedError } from "../../index";
 
 /**
  * Validates an AuthSessionId and returns a [ScimUser]{@link ScimUser}
  *
  * @param {string} systemBaseUri SystemBaseUri for the tenant
  * @param {string} authSessionId AuthSessionId which should be validated
- * @returns {ScimUser}
- *
  * @throws {@link UnauthorizedError} indicates an invalid authSessionId or no authSessionId was sent.
  * @throws {@link ForbiddenError} indicates that a valid authSessionId for an external user was sent but external validation was set to false.
  * @throws {@link NotFoundError} indicates that a valid token for a deleted user was sent.
+ *
+ * @category Authentication
  *
  * @example ```typescript
  * const user: ScimUser = await validateAuthSessionId("https://monster-ag.d-velop.cloud", "vXo4FMlnYYiGArNfjfJHEJDNWfjfejglgnewjgrjokgajifahfhdahfuewfhlR/4FxJxmNsjlq2XgWQm/GYVBq/hEvsJy1BK4WLoCXek=&ga8gds7gafkajgkj24ut8ugudash34jGlDG&dr6j0zusHVN8PcyerI0YXqRu30f8AGoUaZ6vInCDtZInS6aK2PplAelsv9t8");
