+++ packages/identityprovider/src/index.ts
@@ -27,11 +27,11 @@ import { followHalJson } from "@dvelop-sdk/axios-hal-json";
 axios.interceptors.request.use(followHalJson);
 
 export { UnauthorizedError, ForbiddenError, NotFoundError, InvalidAppSessionSignatureError } from "./errors";
-export { AuthSession } from "./get-auth-session/auth-session";
-export { getAuthSession } from "./get-auth-session/get-auth-session";
-export { ScimUser } from "./validate-auth-session-id/scim-user";
-export { validateAuthSessionId } from "./validate-auth-session-id/validate-auth-session-id";
-export { getLoginRedirectionUri } from "./get-login-redirection-uri/get-login-redirection-uri";
-export { requestAppSession } from "./request-app-session/request-app-session";
-export { validateAppSessionSignature } from "./validate-app-session-signature/validate-app-session-signature";
-export { getImpersonatedAuthSessionId } from "./get-impersonated-auth-session-id/get-impersonated-auth-session-id";
+export { AuthSession } from "./authentication/get-auth-session/auth-session";
+export { getAuthSession } from "./authentication/get-auth-session/get-auth-session";
+export { ScimUser } from "./authentication/validate-auth-session-id/scim-user";
+export { validateAuthSessionId } from "./authentication/validate-auth-session-id/validate-auth-session-id";
+export { getLoginRedirectionUri } from "./authentication/get-login-redirection-uri/get-login-redirection-uri";
+export { requestAppSession } from "./authentication/request-app-session/request-app-session";
+export { validateAppSessionSignature } from "./authentication/validate-app-session-signature/validate-app-session-signature";
+export { getImpersonatedAuthSessionId } from "./authentication/get-impersonated-auth-session-id/get-impersonated-auth-session-id";
